requirements.txt
python-dotenv
flask == 3.0.3
Flask_SQLAlchemy == 3.1.1
flask-login == 0.6.3
